+++ test/LightQuery.IntegrationTestsServer/User.cs
@@ -5,5 +5,6 @@
         public int Id { get; set; }
         public string UserName { get; set; }
         public string Email { get; set; }
+        public Animal FavoriteAnimal { get; set; }
     }
 }